Create trafficGen
Requirements This java application depends on the Californium CoAP framework. If using Eclipse, one way of setting it up is importing the Californium Eclipse project and adding it to the Trafikgenerator CoAP build path. Meanwhile, Californium needs both the elementconnector Eclipse project and the JUnit 4 library in its build path.